third_party/blink/web_tests/platform/generic/fast/forms/radio/radio-remove-form-attr-expected.txt

Test that removing a form attribute from an input element leaves any existing radio button group in the document in a valid state.

- Verify initial state
PASS rb1.checked is false
PASS rb2.checked is false
PASS rb3.checked is true
PASS rb4.checked is false
PASS rb5.checked is true


- Remove form attr from rb3, verify that rb5 becomes unchecked state
PASS rb1.checked is false
PASS rb2.checked is false
PASS rb3.checked is true
PASS rb4.checked is false
PASS rb5.checked is false
PASS successfullyParsed is true

TEST COMPLETE

 
  
